Update 02-14-2024
* Mupen64Plus demo runs with LLE RDP (many rendering glitches), works well with gliden64 plugin
* Little-endian model now supported and works (required for gliden64 plugin)
* Added glideN64 plugin support (custom version only for cor64)
* Can render with HLE graphics or LLE graphics
* Some fixes for BassSharp
* Boot related fixes
* Better way to manage CSharp defines
* Memory improvements, fixes
* TLB now works
* FPU Fixes
* PIF controller basics are working now such as controller input
* Added some debug functions to the madcat rdp
* Remove hosting the test roms in the repo
* Started work on workbench scripting with IronPython
* Unit test fixes
* a snippet of IPL2 is assembled on the fly required for 6105 booting
* Other fixes, etc.

+++ src/cor64/Mips/BasicBlock.cs
@@ -10,7 +10,7 @@ namespace cor64.Mips
     {
         private List<TRecord> m_RecordSet;
 
-        protected BasicBlock(ulong address)
+        protected BasicBlock(long address)
         {
             Address = address;
             m_RecordSet = new List<TRecord>();
@@ -23,17 +23,20 @@ namespace cor64.Mips
 
         public TRecord GetLastInst() => m_RecordSet[^1];
 
-        public virtual void UndoLastAppend() {
+        public virtual void UndoLastAppend()
+        {
             m_RecordSet.RemoveAt(m_RecordSet.Count - 1);
         }
 
         public int Size => m_RecordSet.Count;
 
-        public ulong Address { get; protected set; }
+        public long Address { get; protected set; }
 
-        public ulong EndAddress {
-            get {
-                return Address + ((ulong)Size * 4);
+        public long EndAddress
+        {
+            get
+            {
+                return Address + (Size * 4);
             }
         }
